// client/src/shared/lib/types.ts
export enum NotificationVariant {
    success = "SUCCESS",
    error = "ERROR",
    warning = "WARNING",
}

export enum NotificationType {
    system = "SYSTEM",
    info = "INFO",
}

export enum StatusAction {
    total_error_disabled = "TOTAL_ERROR_WITH_BLOCK",
    total_error = "TOTAL_ERROR",
    notify_success = NotificationVariant.success,
    notify_error = NotificationVariant.error,
    notify_warning = NotificationVariant.warning,
}

export type Notification = {
    id: string
    title: string | null
    message: string
    type: NotificationType
    variant: NotificationVariant | null
    date: number
    visible: boolean
    timer: any | 0
}

export type AddNotificationPayload = Pick<Notification, "message" | "type"> &
    Partial<Pick<Notification, "title" | "variant">> & {
        expireTime?: number
    }